Add more details to README
Updated the README file to include additional information about the features of the CSEC_HUB and the technology stack used to build it. Also added more detailed instructions for getting started with the platform. Finally, updated the contribution guidelines to provide more clarity on how to contribute.

This commit should help make the README file more informative and user-friendly, and encourage more contributions to the project.